# docs/swagger.yaml
basePath: /
definitions:
  core.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  core.Repository:
    properties:
      id:
        type: integer
      name:
        type: string
      namespace:
        type: string
      scm_id:
        type: integer
      url:
        type: string
    type: object
  coverage.CodeResponse:
    properties:
      blocks:
        items:
          items:
            type: integer
          type: array
        type: array
      code:
        type: string
      filename:
        type: string
      repo:
        $ref: '#/definitions/core.Repository'
    type: object
  coverage.Coverage:
    properties:
      entries:
        items:
          $ref: '#/definitions/coverage.CoverageEntry'
        type: array
      id:
        format: int64
        type: integer
      revision:
        type: string
      timestamp:
        type: string
      trackerID:
        format: int64
        type: integer
    type: object
  coverage.CoverageEntry:
    properties:
      hits:
        type: integer
      lines:
        type: integer
      name:
        type: string
      profiles:
        additionalProperties:
          $ref: '#/definitions/profile.Profile'
        type: object
    type: object
  coverage.CoverageEntryUploadRequest:
    properties:
      entry:
        type: string
      hits:
        type: integer
      lines:
        type: integer
      profiles:
        items:
          $ref: '#/definitions/profile.Profile'
        type: array
    type: object
  coverage.CoverageListResponse:
    properties:
      coverages:
        items:
          $ref: '#/definitions/coverage.CoverageResponse'
        type: array
      repo:
        $ref: '#/definitions/core.Repository'
    type: object
  coverage.CoverageResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/coverage.CoverageEntry'
        type: array
      index:
        type: integer
      revision:
        type: string
      revision_url:
        type: string
      time:
        type: string
    type: object
  coverage.CoverageUploadRequest:
    properties:
      entries:
        items:
          $ref: '#/definitions/coverage.CoverageEntryUploadRequest'
        type: array
      revision:
        type: string
      time:
        type: string
    type: object
  coverage.FileListResponse:
    properties:
      files:
        items:
          $ref: '#/definitions/coverage.FileResponse'
        type: array
      meta:
        $ref: '#/definitions/coverage.MetaResponse'
      repo:
        $ref: '#/definitions/core.Repository'
    type: object
  coverage.FileResponse:
    properties:
      filename:
        type: string
      hits:
        type: integer
      lines:
        type: integer
    type: object
  coverage.MetaResponse:
    properties:
      hits:
        type: integer
      lines:
        type: integer
      revision:
        type: string
      revision_url:
        type: string
      time:
        type: string
    type: object
  profile.Profile:
    properties:
      blocks:
        description: StartLine, EndLine, Count
        items:
          items:
            type: integer
          type: array
        type: array
      filename:
        type: string
      hits:
        type: integer
      lines:
        type: integer
    type: object
...
  server.createCoverageTrackerRequest:
    properties:
      name:
        type: string
      repo_id:
        type: integer
      visibility:
        type: string
    type: object
...
  tracker.PreviewResponse:
    properties:
      series:
        items:
          $ref: '#/definitions/tracker.PreviewSeriesValues'
        type: array
      tracker:
        $ref: '#/definitions/tracker.TrackerResponse'
    type: object
  tracker.PreviewSeriesValues:
    properties:
      series:
        $ref: '#/definitions/tracker.SeriesModel'
      values:
        items:
          $ref: '#/definitions/tracker.ValueModel'
        type: array
    type: object
  tracker.SeriesModel:
    properties:
      config:
        type: string
      data_type:
        type: string
      id:
        type: integer
      name:
        type: string
      tracker_id:
        type: integer
    type: object
  tracker.TrackerModel:
    properties:
      body:
        type: string
      chart_config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      last_updated_at:
        type: string
      name:
        type: string
      owner_id:
        type: integer
      type:
        type: string
      visibility:
        type: string
    type: object
  tracker.TrackerResponse:
    properties:
      body:
        type: string
      chart_config:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      last_updated_at:
        type: string
      like_count:
        type: integer
      liked:
        type: boolean
      name:
        type: string
      owner_id:
        type: integer
      owner_name:
        type: string
      role:
        description: '"" | "owner" | "editor"'
        type: string
      series:
        description: |-
          Series carries preview data (all series plus their latest values) when
          the list is requested with include=preview. Omitted otherwise.
        items:
          $ref: '#/definitions/tracker.PreviewSeriesValues'
        type: array
      type:
        type: string
      visibility:
        description: '"public" | "private"'
        type: string
    type: object
  tracker.ValueModel:
    properties:
      id:
        type: integer
      seriesId:
        type: integer
      time:
        type: string
      value:
        type: number
    type: object
...
info:
  contact: {}
  description: Mora is a code coverage tracker API.
  title: Mora API
  version: "1.0"
paths:
...
  /api/coverages:
    post:
      consumes:
      - application/json
      description: Create a coverage-type tracker linked to a repository. The repository
        must not already have a coverage tracker.
      parameters:
      - description: Coverage tracker information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.createCoverageTrackerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/tracker.TrackerModel'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Create a coverage tracker
      tags:
      - coverage
  /api/coverages/{trackerId}:
    get:
      description: Return coverage history for a tracker (public list endpoint)
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coverage.CoverageListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: List coverages for a tracker
      tags:
      - coverage
    post:
      consumes:
      - application/json
      description: Upload coverage entries for a repository linked to a tracker
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      - description: Coverage data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/coverage.CoverageUploadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/coverage.Coverage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Upload coverage data
      tags:
      - coverage
  /api/coverages/{trackerId}/{id}/{entry}/files:
    get:
      description: Return the file list and metadata for a coverage entry
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      - description: Coverage index
        in: path
        name: id
        required: true
        type: integer
      - description: Entry name
        in: path
        name: entry
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coverage.FileListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: List files in a coverage entry
      tags:
      - coverage
  /api/coverages/{trackerId}/{id}/{entry}/files/{filepath}:
    get:
      description: Return the source code and coverage blocks for a file in an entry
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      - description: Coverage index
        in: path
        name: id
        required: true
        type: integer
      - description: Entry name
        in: path
        name: entry
        required: true
        type: string
      - description: File path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coverage.CodeResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Get a source file
      tags:
      - coverage
  /api/coverages/{trackerId}/preview:
    get:
      description: Return coverage timeline data as virtual series for a tracker
      parameters:
      - description: Tracker ID
        in: path
        name: trackerId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tracker.PreviewResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/core.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/core.ErrorResponse'
      summary: Preview coverage data
      tags:
      - coverage
...
swagger: "2.0"